+++ evm/package.json
@@ -44,7 +44,6 @@
     "@types/node": "^18.14.5",
     "chai": "^4.3.7",
     "dotenv": "^16.3.1",
-    "ethers": "^5.7.2",
     "mocha": "^10.0.0",
     "prettier": "^2.8.7",
     "prettier-plugin-solidity": "^1.1.3",
