+++ evm/ts/tests/00__environment.ts
@@ -19,7 +19,7 @@ import {
     WORMHOLE_GUARDIAN_SET_INDEX,
     WORMHOLE_MESSAGE_FEE,
     mineWait,
-} from "../src/helpers";
+} from "../src/testing";
 
 describe("Environment", () => {
     const chainNames: ValidNetwork[] = ["Avalanche", "Ethereum", "Base"];
